feat: Extract Radio.Fingerprinting as standalone NuGet package (Phase 5)
Move fingerprinting implementation from Radio.Infrastructure into a
dedicated Radio.Fingerprinting library with clean dependency boundaries.

- Create src/Radio.Fingerprinting project (net10.0, NuGet metadata)
- Move 7 files: FingerprintingOptions, 3 services, 3 repositories
- Introduce IFingerprintDataConnection abstraction to decouple repos
  from FingerprintDbContext (which manages 12 tables, only 3 fingerprinting)
- FingerprintDbContext implements IFingerprintDataConnection interface
- Update ~25 consumer files with new namespace usings
- Add to pack-local.ps1, create package README
- All tests pass (1 pre-existing flaky: WaveformComparisonTests)
