+++ src/Radio.API/Controllers/AudioController.cs
@@ -5,6 +5,7 @@ using Radio.API.Models;
 using Radio.Core.Interfaces.Audio;
 using Radio.Infrastructure.Audio.Fingerprinting;
 using Radio.Infrastructure.Audio.SoundFlow;
+using Radio.Fingerprinting.Services;
 
 namespace Radio.API.Controllers;
