+++ src/Radio.Fingerprinting/GlobalUsings.cs
@@ -1,0 +1,1 @@
+global using Radio.Fingerprinting;
