+++ src/Radio.Fingerprinting/Services/BackgroundIdentificationService.cs
@@ -3,14 +3,13 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using Radio.Core.Configuration;
 using Radio.Core.Events;
 using Radio.Core.Interfaces;
 using Radio.Core.Interfaces.Audio;
 using Radio.Core.Models.Audio;
 using Radio.Metrics;
 
-namespace Radio.Infrastructure.Audio.Fingerprinting;
+namespace Radio.Fingerprinting.Services;
 
 /// <summary>
 /// Background service that periodically identifies audio from active sources using SongRec.
