+++ src/Radio.Infrastructure/Audio/Services/AudioManager.cs
@@ -4,6 +4,7 @@ using Radio.Core.Interfaces.Audio;
 using Radio.Core.Models.Audio;
 using Radio.Infrastructure.Audio.Fingerprinting;
 using Radio.Infrastructure.Audio.SoundFlow;
+using Radio.Fingerprinting.Services;
 
 // Ducking support: AudioManager subscribes to IDuckingService events and applies
 // volume reduction to the active primary source via SoundFlowPlaybackService.
