+++ src/Radio.Infrastructure/Audio/Sources/Primary/USBAudioSourceBase.cs
@@ -6,6 +6,7 @@ using Radio.Core.Models.Audio;
 using Radio.Infrastructure.Audio.Fingerprinting;
 using Radio.Infrastructure.Audio.SoundFlow;
 using Radio.Metrics;
+using Radio.Fingerprinting.Services;
 using SoundFlow.Abstracts.Devices;
 using SoundFlow.Backends.MiniAudio;
 using SoundFlow.Enums;
