+++ src/Radio.Infrastructure/Audio/Sources/Primary/VinylAudioSource.cs
@@ -5,6 +5,7 @@ using Radio.Core.Interfaces.Audio;
 using Radio.Core.Models.Audio;
 using Radio.Infrastructure.Audio.Fingerprinting;
 using Radio.Infrastructure.Audio.SoundFlow;
+using Radio.Fingerprinting.Services;
 
 namespace Radio.Infrastructure.Audio.Sources.Primary;
