+++ tests/Radio.Infrastructure.Tests/Audio/Fingerprinting/BackgroundIdentificationServiceSongChangeTests.cs
@@ -6,6 +6,8 @@ using Radio.Core.Configuration;
 using Radio.Core.Events;
 using Radio.Core.Models.Audio;
 using Radio.Infrastructure.Audio.Fingerprinting;
+using Radio.Fingerprinting.Services;
+using Radio.Fingerprinting;
 
 namespace Radio.Infrastructure.Tests.Audio.Fingerprinting;
