+++ tests/Radio.Infrastructure.Tests/Audio/Fingerprinting/BackgroundIdentificationServiceStatusTests.cs
@@ -5,6 +5,8 @@ using Moq;
 using Radio.Core.Configuration;
 using Radio.Core.Models.Audio;
 using Radio.Infrastructure.Audio.Fingerprinting;
+using Radio.Fingerprinting.Services;
+using Radio.Fingerprinting;
 
 namespace Radio.Infrastructure.Tests.Audio.Fingerprinting;
