+++ tests/Radio.Infrastructure.Tests/Audio/Fingerprinting/MetadataLookupServiceTests.cs
@@ -4,6 +4,8 @@ using Moq;
 using Radio.Core.Configuration;
 using Radio.Core.Interfaces;
 using Radio.Infrastructure.Audio.Fingerprinting;
+using Radio.Fingerprinting.Services;
+using Radio.Fingerprinting;
 
 namespace Radio.Infrastructure.Tests.Audio.Fingerprinting;
