+++ tests/Radio.IntegrationTests/Fingerprinting/FingerprintingConfigurationIntegrationTests.cs
@@ -5,6 +5,9 @@ using Moq;
 using Radio.Core.Configuration;
 using Radio.Core.Models.Audio;
 using Radio.Infrastructure.Audio.Fingerprinting.Data;
+using Radio.Fingerprinting;
+using Radio.Fingerprinting.Abstractions;
+using Radio.Fingerprinting.Data;
 using Radio.IntegrationTests.TestSupport;
 
 namespace Radio.IntegrationTests.Fingerprinting;
