+++ tests/Radio.IntegrationTests/Radio.IntegrationTests.csproj
@@ -24,6 +24,7 @@
   <ItemGroup>
     <ProjectReference Include="..\..\src\Radio.API\Radio.API.csproj" />
     <ProjectReference Include="..\..\src\Radio.Configuration\Radio.Configuration.csproj" />
+    <ProjectReference Include="..\..\src\Radio.Fingerprinting\Radio.Fingerprinting.csproj" />
     <ProjectReference Include="..\..\src\Radio.Infrastructure\Radio.Infrastructure.csproj" />
     <ProjectReference Include="..\..\src\Radio.Core\Radio.Core.csproj" />
     <ProjectReference Include="..\..\src\RTLSDRCore\RTLSDRCore.csproj" />
